+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 2.5.0
 - Added "gitGrace.stashClear" command.
 - Amended progress location back to the status bar.
+- Amended "gitGrace.fetch" and "gitGrace.checkout" commands so they will not ask for a rebase/merge if the remote commits were solely amended.
 
 ### 2.4.0
 - Amended "gitGrace.fetch" and "gitGrace.checkout" commands so they ask if users want to sync the local branch with its remote branch.